Fix discrepancies
Label/goto numbers are `unsigned int`s, but some of the functions that use them were taking `uint8_t`s.

+++ src/parse.h
@@ -23,7 +23,7 @@ enum {
     TYPE_MASK_U24
 };
 
-bool JumpForward(uint8_t *, uint8_t *, unsigned int, uint8_t, uint8_t);
+bool JumpForward(uint8_t *, uint8_t *, unsigned int, unsigned int, unsigned int);
 
 bool JumpBackwards(uint8_t *, uint8_t);
